requirements.txt
pyneurosdk2==1.0.15
PyQt6>=6.6
pygame>=2.5
numpy>=1.25
scipy>=1.11
scikit-learn>=1.3
SpeechRecognition>=3.10
pyttsx3>=2.90
pynput>=1.7
platformdirs>=3.11
